fix: page path name (#162)
