+++ app/admin/category/_ui/shared/_api/upload-file-with-presigned-url.ts
