+++ app/admin/category/_ui/shared/file-input/index.tsx
