+++ app/admin/category/_ui/shared/manage-table/file-input/index.tsx
