+++ app/admin/category/_ui/shared/manage-table/styles.module.scss
