+++ app/admin/category/_ui/shared/manage-table/utils/add-index-to-data.tsx
