+++ app/admin/category/_ui/stock/stock-manage/_api/get-stocks.ts
