+++ app/admin/category/_ui/stock/stock-manage/_api/toggle-stock-active-state.ts
