+++ app/admin/category/_ui/stock/stock-manage/_hooks/query/use-stocks-data.ts
