+++ app/admin/category/_ui/stock/stock-manage/_ui/active-stock-manage-table.tsx
