+++ app/admin/category/_ui/stock/stock-manage/_ui/inactive-stock-manage-table.tsx
