+++ app/admin/category/_ui/stock/stock-manage/_ui/stock-active-state-toggle-button.tsx
