+++ app/admin/category/_ui/stock/stock-manage/_ui/stock-post-button/styles.module.scss
