+++ app/admin/category/_ui/stock/stock-manage/index.tsx
