+++ app/admin/category/_ui/stock/stock-manage/styles.module.scss
