+++ app/admin/category/_ui/stock/stock-manage/types.ts
