+++ app/admin/category/_ui/trade/trade-manage/_api/toggle-trade-active-state.ts
