+++ app/admin/category/_ui/trade/trade-manage/_hooks/query/use-trades-data.ts
