+++ app/admin/category/_ui/trade/trade-manage/_ui/inactive-trade-manage-table.tsx
