+++ app/admin/category/_ui/trade/trade-manage/index.tsx
